requirements.txt
streamlit
groq
python-dotenv
pandas
numpy
pydeck